+++ src/search/heuristics/cea_heuristic.cc
@@ -2,7 +2,7 @@
 
 #include "domain_transition_graph.h"
 
-#include "../evaluators/default_value_axioms_evaluator.h"
+#include "../evaluators/axiom_handling_evaluator.h"
 #include "../plugins/plugin.h"
 #include "../task_utils/task_properties.h"
 #include "../utils/logging.h"
