+++ src/search/heuristics/cg_heuristic.cc
@@ -3,7 +3,7 @@
 #include "cg_cache.h"
 #include "domain_transition_graph.h"
 
-#include "../evaluators/default_value_axioms_evaluator.h"
+#include "../evaluators/axiom_handling_evaluator.h"
 #include "../plugins/plugin.h"
 #include "../task_utils/task_properties.h"
 #include "../utils/logging.h"
