+++ src/search/heuristics/max_heuristic.cc
@@ -1,6 +1,6 @@
 #include "max_heuristic.h"
 
-#include "../evaluators/default_value_axioms_evaluator.h"
+#include "../evaluators/axiom_handling_evaluator.h"
 #include "../plugins/plugin.h"
 #include "../utils/logging.h"
