+++ src/search/landmarks/landmark_sum_heuristic.cc
@@ -5,7 +5,7 @@
 #include "landmark_status_manager.h"
 #include "util.h"
 
-#include "../evaluators/default_value_axioms_evaluator.h"
+#include "../evaluators/axiom_handling_evaluator.h"
 #include "../plugins/plugin.h"
 #include "../task_utils/successor_generator.h"
 #include "../task_utils/task_properties.h"
